+++ ostium_python_sdk/ostium.py
@@ -16,8 +16,9 @@ class OpenOrderType(Enum):
 
 
 class Ostium:
-    def __init__(self, w3: Web3, usdc_address: str, ostium_trading_storage_address: str, ostium_trading_address: str, private_key: str) -> None:
+    def __init__(self, w3: Web3, usdc_address: str, ostium_trading_storage_address: str, ostium_trading_address: str, private_key: str, verbose=False) -> None:
         self.web3 = w3
+        self.verbose = verbose
         self.private_key = private_key
         self.usdc_address = usdc_address
         self.ostium_trading_storage_address = ostium_trading_storage_address
@@ -32,6 +33,10 @@ class Ostium:
             address=self.ostium_trading_address, abi=ostium_trading_abi)
 
         self.slippage_percentage = 2  # 2%
+
+    def log(self, message):
+        if self.verbose:
+            print(message)
 
     def set_slippage_percentage(self, slippage_percentage):
         self.slippage_percentage = slippage_percentage
@@ -60,12 +65,13 @@ class Ostium:
                 "Private key is required for Ostium platform write-operations")
 
     def perform_trade(self, trade_params, at_price):
+        self.log(f"Performing trade with params: {trade_params}")
         account = self._get_account()
         amount = to_base_units(trade_params['collateral'], decimals=6)
         self.__approve(account, amount)
 
         try:
-            print("Final trade parameters being sent:", trade_params)
+            self.log(f"Final trade parameters being sent: {trade_params}")
             tp_price, sl_price = get_tp_sl_prices(trade_params)
 
             trade = {
@@ -102,50 +108,52 @@ class Ostium:
                 signed_tx.raw_transaction)
             trade_receipt = self.web3.eth.wait_for_transaction_receipt(
                 trade_tx_hash)
-            # print('Trade Receipt:', trade_receipt)
+            self.log(f"Trade Receipt: {trade_receipt}")
             return trade_receipt
 
         except Exception as e:
-            reason_string, suggestion = fromErrorCodeToMessage(e)
+            reason_string, suggestion = fromErrorCodeToMessage(
+                e, verbose=self.verbose)
             print(
                 f"An error ({str(e)}) occurred during the trading process - parsed as {reason_string}")
             raise Exception(
                 f'{reason_string}\n\n{suggestion}' if suggestion != None else reason_string)
 
-    def cancel_limit_order(self, pair_id, index):
+    def cancel_limit_order(self, pair_id, trade_index):
         account = self._get_account()
 
         trade_tx = self.ostium_trading_contract.functions.cancelOpenLimitOrder(
-            int(pair_id), int(index)).build_transaction({'from': account.address})
+            int(pair_id), int(trade_index)).build_transaction({'from': account.address})
         trade_tx['nonce'] = self.get_nonce(account.address)
 
         signed_tx = self.web3.eth.account.sign_transaction(
             trade_tx, private_key=self.private_key)
         trade_tx_hash = self.web3.eth.send_raw_transaction(
             signed_tx.raw_transaction)
-        print('Cancel Limit Order TX Hash:', trade_tx_hash.hex())
+        self.log(f"Cancel Limit Order TX Hash: {trade_tx_hash.hex()}")
 
         trade_receipt = self.web3.eth.wait_for_transaction_receipt(
             trade_tx_hash)
-        print('Cancel Limit Order Receipt:', trade_receipt)
+        self.log(f"Cancel Limit Order Receipt: {trade_receipt}")
         return trade_receipt
 
-    def close_trade(self, pairID, index):
+    def close_trade(self, pair_id, trade_index):
+        self.log(f"Closing trade for pair {pair_id}, index {trade_index}")
         account = self._get_account()
 
         trade_tx = self.ostium_trading_contract.functions.closeTradeMarket(
-            int(pairID), int(index)).build_transaction({'from': account.address})
+            int(pair_id), int(trade_index)).build_transaction({'from': account.address})
         trade_tx['nonce'] = self.get_nonce(account.address)
 
         signed_tx = self.web3.eth.account.sign_transaction(
             trade_tx, private_key=self.private_key)
         trade_tx_hash = self.web3.eth.send_raw_transaction(
             signed_tx.raw_transaction)
-        print('Trade TX Hash:', trade_tx_hash.hex())
+        self.log(f"Trade TX Hash: {trade_tx_hash.hex()}")
 
         trade_receipt = self.web3.eth.wait_for_transaction_receipt(
             trade_tx_hash)
-        # print('Trade Receipt:', trade_receipt)
+        self.log(f"Trade Receipt: {trade_receipt}")
         return trade_receipt
 
     def add_collateral(self, pairID, index, collateral):
@@ -162,11 +170,11 @@ class Ostium:
                 add_collateral_tx, private_key=self.private_key)
             add_collateral_tx_hash = self.web3.eth.send_raw_transaction(
                 signed_tx.raw_transaction)
-            print('Add Collateral TX Hash:', add_collateral_tx_hash.hex())
+            self.log(f"Add Collateral TX Hash: {add_collateral_tx_hash.hex()}")
 
             add_collateral_receipt = self.web3.eth.wait_for_transaction_receipt(
                 add_collateral_tx_hash)
-            print('Add Collateral Receipt:', add_collateral_receipt)
+            self.log(f"Add Collateral Receipt: {add_collateral_receipt}")
             return add_collateral_receipt
 
         except Exception as e:
@@ -174,20 +182,22 @@ class Ostium:
             traceback.print_exc()
             raise e
 
-    def update_tp(self, pairID, index, tp):
+    def update_tp(self, pair_id, trade_index, tp_price):
+        self.log(
+            f"Updating TP for pair {pair_id}, index {trade_index} to {tp_price}")
         account = self._get_account()
         try:
-            tp_value = to_base_units(tp, decimals=18)
+            tp_value = to_base_units(tp_price, decimals=18)
 
             update_tp_tx = self.ostium_trading_contract.functions.updateTp(
-                int(pairID), int(index), tp_value).build_transaction({'from': account.address})
+                int(pair_id), int(trade_index), tp_value).build_transaction({'from': account.address})
             update_tp_tx['nonce'] = self.get_nonce(account.address)
 
             signed_tx = self.web3.eth.account.sign_transaction(
                 update_tp_tx, private_key=self.private_key)
             update_tp_tx_hash = self.web3.eth.send_raw_transaction(
                 signed_tx.raw_transaction)
-            print('Update TP TX Hash:', update_tp_tx_hash.hex())
+            self.log(f"Update TP TX Hash: {update_tp_tx_hash.hex()}")
 
         except Exception as e:
             print("An error occurred during the update tp process:")
@@ -207,10 +217,11 @@ class Ostium:
                 update_sl_tx, private_key=self.private_key)
             update_sl_tx_hash = self.web3.eth.send_raw_transaction(
                 signed_tx.raw_transaction)
-            print('Update SL TX Hash:', update_sl_tx_hash.hex())
+            self.log(f"Update SL TX Hash: {update_sl_tx_hash.hex()}")
 
         except Exception as e:
-            reason_string, suggestion = fromErrorCodeToMessage(str(e))
+            reason_string, suggestion = fromErrorCodeToMessage(
+                str(e), verbose=self.verbose)
             print(
                 f"An error occurred during the update sl process: {reason_string}")
             raise Exception(
@@ -232,11 +243,11 @@ class Ostium:
                 approve_tx, private_key=self.private_key)
             approve_tx_hash = self.web3.eth.send_raw_transaction(
                 signed_tx.raw_transaction)
-            print('Approval TX Hash:', approve_tx_hash.hex())
+            self.log(f"Approval TX Hash: {approve_tx_hash.hex()}")
 
             approve_receipt = self.web3.eth.wait_for_transaction_receipt(
                 approve_tx_hash)
-            print('Approval Receipt:', approve_receipt)
+            self.log(f"Approval Receipt: {approve_receipt}")
 
     def withdraw(self, amount, receiving_address):
         account = self._get_account()
@@ -258,15 +269,16 @@ class Ostium:
                 transfer_tx, private_key=self.private_key)
             transfer_tx_hash = self.web3.eth.send_raw_transaction(
                 signed_tx.raw_transaction)
-            print('Transfer TX Hash:', transfer_tx_hash.hex())
+            self.log(f"Transfer TX Hash: {transfer_tx_hash.hex()}")
 
             transfer_receipt = self.web3.eth.wait_for_transaction_receipt(
                 transfer_tx_hash)
-            print('Transfer Receipt:', transfer_receipt)
+            self.log(f"Transfer Receipt: {transfer_receipt}")
             return transfer_receipt
 
         except Exception as e:
-            reason_string, suggestion = fromErrorCodeToMessage(str(e))
+            reason_string, suggestion = fromErrorCodeToMessage(
+                str(e), verbose=self.verbose)
             print(
                 f"An error occurred during the transfer process: {reason_string}")
             raise Exception(
@@ -274,10 +286,7 @@ class Ostium:
 
     def update_limit_order(self, pair_id, index, pvt_key, price=None, tp=None, sl=None):
         try:
-            print('update_limit_order called with pair_id', pair_id, 'index',
-                  index, 'pvt_key', pvt_key, 'price', price, 'tp', tp, 'sl', sl)
             account = self.web3.eth.account.from_key(pvt_key)
-
             # Get existing order details (tbd why read from storage)
             existing_order = self.ostium_trading_storage_contract.functions.getOpenLimitOrder(
                 account.address,
@@ -285,7 +294,7 @@ class Ostium:
                 int(index)
             ).call()
 
-            print('existing_order', existing_order)
+            self.log(f"existing_order {existing_order}")
             # Use existing values if new values are not provided
             price_value = convert_to_scaled_integer(
                 price) if price is not None else existing_order[1]  # openPrice
@@ -294,8 +303,6 @@ class Ostium:
             sl_value = convert_to_scaled_integer(
                 sl) if sl is not None else existing_order[3]    # sl
 
-            print('calling updateOpenLimitOrder with price_value',
-                  price_value, 'tp_value', tp_value, 'sl_value', sl_value)
             trade_tx = self.ostium_trading_contract.functions.updateOpenLimitOrder(
                 int(pair_id),
                 int(index),
@@ -310,15 +317,16 @@ class Ostium:
                 trade_tx, private_key=account.key)
             trade_tx_hash = self.web3.eth.send_raw_transaction(
                 signed_tx.raw_transaction)
-            print('Update Limit Order TX Hash:', trade_tx_hash.hex())
+            self.log(f"Update Limit Order TX Hash: {trade_tx_hash.hex()}")
 
             trade_receipt = self.web3.eth.wait_for_transaction_receipt(
                 trade_tx_hash)
-            print('Update Limit Order Receipt:', trade_receipt)
+            self.log(f"Update Limit Order Receipt: {trade_receipt}")
             return trade_receipt
 
         except Exception as e:
-            reason_string, suggestion = fromErrorCodeToMessage(str(e))
+            reason_string, suggestion = fromErrorCodeToMessage(
+                str(e), verbose=self.verbose)
             print(
                 f"An error occurred during the update limit order process: {reason_string}")
             raise Exception(