+++ ostium_python_sdk/price.py
@@ -3,8 +3,13 @@ from typing import Tuple
 
 
 class Price:
-    def __init__(self):
+    def __init__(self, verbose=False):
+        self.verbose = verbose
         self.base_url = "https://listener.ostium.io"
+
+    def log(self, message):
+        if self.verbose:
+            print(message)
 
     # Returns a list of price data, e.g: [{'feed_id': '0x00039d9e45394f473ab1f050a1b963e6b05351e52d71e507509ada0c95ed75b8', 'bid': 97241.43864211132, 'mid': 97243.36503172085, 'ask': 97245.2739217016, 'isMarketOpen': True, 'from': 'BTC', 'to': 'USD', 'timestampSeconds': 1740043714}, ...]
     async def get_latest_prices(self):
@@ -26,23 +31,17 @@ class Price:
         for price_data in prices:
             if (price_data.get('from') == from_asset and
                     price_data.get('to') == to_asset):
-                print(f"get_latest_price_json: {price_data}")
+                self.log(f"get_latest_price_json: {price_data}")
                 return price_data
         raise ValueError(f"No price found for pair: {from_asset}/{to_asset}")
 
     # Returns a mid price and isMarketOpen tuple, e.g: (97243.36503172085, True)
-    async def get_price(self, from_asset: str, to_asset: str) -> Tuple[float, bool]:
-        """
-        Get the latest price and market status for a specific asset pair.
-        Args:
-            from_asset: The base asset symbol (e.g., "BTC")
-            to_asset: The quote asset symbol (e.g., "USD")
-        Returns:
-            Tuple[float, bool]: The latest price and market open status for the asset pair
-        """
+    async def get_price(self, from_currency, to_currency):
+        self.log(f"Getting price for {from_currency}/{to_currency}")
         prices = await self.get_latest_prices()
         for price_data in prices:
-            if (price_data.get('from') == from_asset and
-                    price_data.get('to') == to_asset):
+            if (price_data.get('from') == from_currency and
+                    price_data.get('to') == to_currency):
                 return float(price_data.get('mid', 0)), price_data.get('isMarketOpen', False)
-        raise ValueError(f"No price found for pair: {from_asset}/{to_asset}")
+        raise ValueError(
+            f"No price found for pair: {from_currency}/{to_currency}")